requirements.txt
streamlit==1.54.0
pandas==2.3.3
faker==40.4.0
mysql-connector-python==9.6.0
numpy==2.4.2
pyarrow==23.0.1
python-dotenv==1.0.0
